ourapp/requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-decouple==3.8

# Server
gunicorn==21.2.0
whitenoise==6.5.0

# Database - SQLite is included with Django and works out of the box
# For PostgreSQL (optional - comment out if not needed):
# psycopg2-binary==2.9.7

# File monitoring dependencies
watchdog==3.0.0

# Web security dependencies
jsonfield==3.1.0
python-whois==0.8.0
beautifulsoup4==4.12.2
requests==2.31.0

# Selenium for screenshots (chrome/selenium setup required)
selenium==4.15.2
webdriver-manager==4.0.1

# HTML processing
# lxml==4.9.3
html5lib==1.1

# Security and analysis
cryptography==41.0.5

# Utilities
PyYAML==6.0.1
python-dateutil==2.8.2